NEW: Fix share deleting snag for external upload shares [q11867]
git-svn-id: http://svn.resourcespace.com/svn/rs/branches/20201120_nbraddy_shareuploadlink_q11867@16923 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ pages/manage_external_shares.php
@@ -42,7 +42,7 @@ if($delete_access_key != "" && enforcePostRequest($ajax))
         'success' => false
     );
 
-    if($deleteresource != "")
+    if($deleteresource != "" && $deleteresource != "-")
         {
         delete_resource_access_key($deleteresource, $delete_access_key);
         $response['success'] = true;
